+++ packages/e2e/local-network/scripts/pools/update-node-utils.sh
@@ -9,22 +9,26 @@ cd "$root"
 
 export PATH=$PWD/bin:$PATH
 
+SP_NODE_ID="$1"
+
+mkdir ${SP_NODE_ID}
+
 clean() {
-  rm -rf pool-owner-registration.cert pool-owner-delegation.cert wallets-tx.raw wallets-tx.signed fullUtxo.out balance.out params.json stake.cert pool.cert deleg.cert tx.tmp tx.raw tx.signed
+  rm -rf ${SP_NODE_ID}
 }
 
 getAddressBalance() {
   cardano-cli query utxo \
     --address "$1" \
-    --testnet-magic 888 >fullUtxo.out
+    --testnet-magic 888 > ${SP_NODE_ID}/fullUtxo.out
 
-  tail -n +3 fullUtxo.out | sort -k3 -nr >balance.out
+  tail -n +3 ${SP_NODE_ID}/fullUtxo.out | sort -k3 -nr > ${SP_NODE_ID}/balance.out
 
   total_balance=0
   while read -r utxo; do
     utxo_balance=$(awk '{ print $3 }' <<<"${utxo}")
     total_balance=$(("$total_balance" + "$utxo_balance"))
-  done <balance.out
+  done < ${SP_NODE_ID}/balance.out
 
   echo ${total_balance}
 }
@@ -53,10 +57,10 @@ updatePool() {
   fi
 
   # get the protocol parameters
-  cardano-cli query protocol-parameters --testnet-magic 888 --out-file params.json
+  cardano-cli query protocol-parameters --testnet-magic 888 --out-file ${SP_NODE_ID}/params.json
 
-  genesisVKey=network-files/utxo-keys/utxo3.vkey
-  genesisSKey=network-files/utxo-keys/utxo3.skey
+  genesisVKey=network-files/utxo-keys/utxo${SP_NODE_ID}.vkey
+  genesisSKey=network-files/utxo-keys/utxo${SP_NODE_ID}.skey
   genesisAddr=$(cardano-cli address build --payment-verification-key-file "$genesisVKey" --testnet-magic 888)
 
   stakeVKey=network-files/pools/staking-reward"${SP_NODE_ID}".vkey
@@ -68,6 +72,8 @@ updatePool() {
   delegatorStakeKey=network-files/stake-delegator-keys/staking"${SP_NODE_ID}".vkey
   delegatorPaymentSKey=network-files/stake-delegator-keys/payment"${SP_NODE_ID}".skey
   delegatorStakeSKey=network-files/stake-delegator-keys/staking"${SP_NODE_ID}".skey
+
+  keyDeposit=2000000
 
   POOL_ID=$(cardano-cli stake-pool id --cold-verification-key-file "$coldVKey" --output-format "hex")
 
@@ -82,15 +88,15 @@ updatePool() {
     --tx-in "$utxo" \
     --tx-out "$stakeAddr"+"$POOL_OWNER_STAKE" \
     --testnet-magic 888 \
-    --out-file wallets-tx.raw 2>&1
+    --out-file ${SP_NODE_ID}/wallets-tx.raw 2>&1
 
   cardano-cli transaction sign \
-    --tx-body-file wallets-tx.raw \
+    --tx-body-file ${SP_NODE_ID}/wallets-tx.raw \
     --signing-key-file "$genesisSKey" \
     --testnet-magic 888 \
-    --out-file wallets-tx.signed 2>&1
+    --out-file ${SP_NODE_ID}/wallets-tx.signed 2>&1
 
-  cardano-cli transaction submit --testnet-magic 888 --tx-file wallets-tx.signed 2>&1
+  cardano-cli transaction submit --testnet-magic 888 --tx-file ${SP_NODE_ID}/wallets-tx.signed 2>&1
 
   updatedBalance=$(getAddressBalance "$stakeAddr")
 
@@ -103,7 +109,7 @@ updatePool() {
   currentBalance=$(getAddressBalance "$genesisAddr")
   cardano-cli stake-address registration-certificate \
     --stake-verification-key-file "$stakeVKey" \
-    --out-file pool-owner-registration.cert
+    --out-file ${SP_NODE_ID}/pool-owner-registration.cert
 
   utxo=$(cardano-cli query utxo --address "$genesisAddr" --testnet-magic 888 | awk 'NR == 3 {printf("%s#%s", $1, $2)}')
 
@@ -112,38 +118,38 @@ updatePool() {
     --tx-out "$genesisAddr"+0 \
     --invalid-hereafter 5000000 \
     --fee 0 \
-    --out-file tx.tmp \
-    --certificate pool-owner-registration.cert
+    --out-file ${SP_NODE_ID}/tx.tmp \
+    --certificate ${SP_NODE_ID}/pool-owner-registration.cert
 
   fee=$(cardano-cli transaction calculate-min-fee \
-    --tx-body-file tx.tmp \
+    --tx-body-file ${SP_NODE_ID}/tx.tmp \
     --tx-in-count 1 \
     --tx-out-count 1 \
     --testnet-magic 888 \
     --witness-count 2 \
     --byron-witness-count 0 \
-    --protocol-params-file ./params.json | awk '{ print $1 }')
+    --protocol-params-file ${SP_NODE_ID}/params.json | awk '{ print $1 }')
 
   initialBalance=$(getAddressBalance "$genesisAddr")
-  txOut=$((initialBalance - fee))
+  txOut=$((initialBalance - fee - keyDeposit))
 
   cardano-cli transaction build-raw \
     --tx-in "$utxo" \
     --tx-out "$genesisAddr"+"$txOut" \
     --invalid-hereafter 5000000 \
     --fee "$fee" \
-    --certificate pool-owner-registration.cert \
-    --out-file tx.raw
+    --certificate ${SP_NODE_ID}/pool-owner-registration.cert \
+    --out-file ${SP_NODE_ID}/tx.raw
 
   cardano-cli transaction sign \
-    --tx-body-file tx.raw \
+    --tx-body-file ${SP_NODE_ID}/tx.raw \
     --signing-key-file "$genesisSKey" \
     --signing-key-file "$stakeKey" \
     --testnet-magic 888 \
-    --out-file tx.signed
+    --out-file ${SP_NODE_ID}/tx.signed
 
   cardano-cli transaction submit \
-    --tx-file tx.signed \
+    --tx-file ${SP_NODE_ID}/tx.signed \
     --testnet-magic 888
 
   updatedBalance=$(getAddressBalance "$genesisAddr")
@@ -158,7 +164,7 @@ updatePool() {
   cardano-cli stake-address delegation-certificate \
     --stake-verification-key-file "$stakeVKey" \
     --cold-verification-key-file "$coldVKey" \
-    --out-file pool-owner-delegation.cert
+    --out-file ${SP_NODE_ID}/pool-owner-delegation.cert
 
   utxo=$(cardano-cli query utxo --address "$genesisAddr" --testnet-magic 888 | awk 'NR == 3 {printf("%s#%s", $1, $2)}')
 
@@ -167,17 +173,17 @@ updatePool() {
     --tx-out "$genesisAddr"+0 \
     --invalid-hereafter 5000000 \
     --fee 0 \
-    --out-file tx.tmp \
-    --certificate pool-owner-delegation.cert
+    --out-file ${SP_NODE_ID}/tx.tmp \
+    --certificate ${SP_NODE_ID}/pool-owner-delegation.cert
 
   fee=$(cardano-cli transaction calculate-min-fee \
-    --tx-body-file tx.tmp \
+    --tx-body-file ${SP_NODE_ID}/tx.tmp \
     --tx-in-count 1 \
     --tx-out-count 1 \
     --testnet-magic 888 \
     --witness-count 2 \
     --byron-witness-count 0 \
-    --protocol-params-file ./params.json | awk '{ print $1 }')
+    --protocol-params-file ${SP_NODE_ID}/params.json | awk '{ print $1 }')
 
   initialBalance=$(getAddressBalance "$genesisAddr")
   txOut=$((initialBalance - fee))
@@ -187,18 +193,18 @@ updatePool() {
     --tx-out "$genesisAddr"+"$txOut" \
     --invalid-hereafter 5000000 \
     --fee "$fee" \
-    --certificate pool-owner-delegation.cert \
-    --out-file tx.raw
+    --certificate ${SP_NODE_ID}/pool-owner-delegation.cert \
+    --out-file ${SP_NODE_ID}/tx.raw
 
   cardano-cli transaction sign \
-    --tx-body-file tx.raw \
+    --tx-body-file ${SP_NODE_ID}/tx.raw \
     --signing-key-file "$genesisSKey" \
     --signing-key-file "$stakeKey" \
     --testnet-magic 888 \
-    --out-file tx.signed
+    --out-file ${SP_NODE_ID}/tx.signed
 
   cardano-cli transaction submit \
-    --tx-file tx.signed \
+    --tx-file ${SP_NODE_ID}/tx.signed \
     --testnet-magic 888
 
   updatedBalance=$(getAddressBalance "$genesisAddr")
@@ -218,7 +224,7 @@ updatePool() {
   cardano-cli stake-address delegation-certificate \
     --stake-verification-key-file "$delegatorStakeKey" \
     --stake-pool-id "$POOL_ID" \
-    --out-file deleg.cert
+    --out-file ${SP_NODE_ID}/deleg.cert
 
   utxo=$(cardano-cli query utxo --address "$paymentAddr" --testnet-magic 888 | awk 'NR == 3 {printf("%s#%s", $1, $2)}')
 
@@ -227,17 +233,17 @@ updatePool() {
     --tx-out "$paymentAddr"+0 \
     --invalid-hereafter 5000000 \
     --fee 0 \
-    --out-file tx.tmp \
-    --certificate deleg.cert
+    --out-file ${SP_NODE_ID}/tx.tmp \
+    --certificate ${SP_NODE_ID}/deleg.cert
 
   fee=$(cardano-cli transaction calculate-min-fee \
-    --tx-body-file tx.tmp \
+    --tx-body-file ${SP_NODE_ID}/tx.tmp \
     --tx-in-count 1 \
     --tx-out-count 1 \
     --testnet-magic 888 \
     --witness-count 2 \
     --byron-witness-count 0 \
-    --protocol-params-file ./params.json | awk '{ print $1 }')
+    --protocol-params-file ${SP_NODE_ID}/params.json | awk '{ print $1 }')
 
   initialBalance=$(getAddressBalance "$paymentAddr")
   txOut=$((initialBalance - fee))
@@ -247,18 +253,18 @@ updatePool() {
     --tx-out "$paymentAddr"+"$txOut" \
     --invalid-hereafter 5000000 \
     --fee "$fee" \
-    --certificate-file deleg.cert \
-    --out-file tx.raw
+    --certificate-file ${SP_NODE_ID}/deleg.cert \
+    --out-file ${SP_NODE_ID}/tx.raw
 
   cardano-cli transaction sign \
-    --tx-body-file tx.raw \
+    --tx-body-file ${SP_NODE_ID}/tx.raw \
     --signing-key-file "$delegatorPaymentSKey" \
     --signing-key-file "$delegatorStakeSKey" \
     --testnet-magic 888 \
-    --out-file tx.signed
+    --out-file ${SP_NODE_ID}/tx.signed
 
   cardano-cli transaction submit \
-    --tx-file tx.signed \
+    --tx-file ${SP_NODE_ID}/tx.signed \
     --testnet-magic 888
 
   updatedBalance=$(getAddressBalance "$paymentAddr")
@@ -287,7 +293,7 @@ updatePool() {
       --pool-relay-port 300"$SP_NODE_ID" \
       --metadata-url "${METADATA_URL}" \
       --metadata-hash "${METADATA_HASH}" \
-      --out-file pool.cert
+      --out-file ${SP_NODE_ID}/pool.cert
   else
     cardano-cli stake-pool registration-certificate \
       --cold-verification-key-file "$coldVKey" \
@@ -300,7 +306,7 @@ updatePool() {
       --testnet-magic 888 \
       --pool-relay-ipv4 127.0.0.1 \
       --pool-relay-port 300"$SP_NODE_ID" \
-      --out-file pool.cert
+      --out-file ${SP_NODE_ID}/pool.cert
   fi
 
   utxo=$(cardano-cli query utxo --address "$paymentAddr" --testnet-magic 888 | awk 'NR == 3 {printf("%s#%s", $1, $2)}')
@@ -310,17 +316,17 @@ updatePool() {
     --tx-out "$paymentAddr"+"$txOut" \
     --invalid-hereafter 500000 \
     --fee 0 \
-    --certificate-file pool.cert \
-    --out-file tx.tmp
+    --certificate-file ${SP_NODE_ID}/pool.cert \
+    --out-file ${SP_NODE_ID}/tx.tmp
 
   fee=$(cardano-cli transaction calculate-min-fee \
-    --tx-body-file tx.tmp \
+    --tx-body-file ${SP_NODE_ID}/tx.tmp \
     --tx-in-count 1 \
     --tx-out-count 1 \
     --testnet-magic 888 \
     --witness-count 3 \
     --byron-witness-count 0 \
-    --protocol-params-file ./params.json | awk '{ print $1 }')
+    --protocol-params-file ${SP_NODE_ID}/params.json | awk '{ print $1 }')
 
   initialBalance=$(getAddressBalance "$paymentAddr")
   txOut=$((initialBalance - fee))
@@ -330,19 +336,19 @@ updatePool() {
     --tx-out "$paymentAddr"+"$txOut" \
     --invalid-hereafter 500000 \
     --fee "$fee" \
-    --certificate-file pool.cert \
-    --out-file tx.raw
+    --certificate-file ${SP_NODE_ID}/pool.cert \
+    --out-file ${SP_NODE_ID}/tx.raw
 
   cardano-cli transaction sign \
-    --tx-body-file tx.raw \
+    --tx-body-file ${SP_NODE_ID}/tx.raw \
     --signing-key-file "$delegatorPaymentSKey" \
     --signing-key-file "$coldKey" \
     --signing-key-file "$stakeKey" \
     --testnet-magic 888 \
-    --out-file tx.signed
+    --out-file ${SP_NODE_ID}/tx.signed
 
   cardano-cli transaction submit \
-    --tx-file tx.signed \
+    --tx-file ${SP_NODE_ID}/tx.signed \
     --testnet-magic 888
 
   updatedBalance=$(getAddressBalance "$paymentAddr")
@@ -366,10 +372,10 @@ deregisterPool() {
   done
 
   # get the protocol parameters
-  cardano-cli query protocol-parameters --testnet-magic 888 --out-file params.json
+  cardano-cli query protocol-parameters --testnet-magic 888 --out-file ${SP_NODE_ID}/params.json
 
-  genesisVKey=network-files/utxo-keys/utxo3.vkey
-  genesisSKey=network-files/utxo-keys/utxo3.skey
+  genesisVKey=network-files/utxo-keys/utxo${SP_NODE_ID}.vkey
+  genesisSKey=network-files/utxo-keys/utxo${SP_NODE_ID}.skey
   genesisAddr=$(cardano-cli address build --payment-verification-key-file "$genesisVKey" --testnet-magic 888)
   stakeKey=network-files/pools/staking-reward"${SP_NODE_ID}".skey
   coldVKey=network-files/pools/cold"${SP_NODE_ID}".vkey
@@ -391,7 +397,7 @@ deregisterPool() {
   cardano-cli stake-address delegation-certificate \
     --stake-verification-key-file "$delegatorStakeKey" \
     --stake-pool-id "$POOL_ID" \
-    --out-file deleg.cert
+    --out-file ${SP_NODE_ID}/deleg.cert
 
   utxo=$(cardano-cli query utxo --address "$paymentAddr" --testnet-magic 888 | awk 'NR == 3 {printf("%s#%s", $1, $2)}')
 
@@ -400,17 +406,17 @@ deregisterPool() {
     --tx-out "$paymentAddr"+0 \
     --invalid-hereafter 5000000 \
     --fee 0 \
-    --out-file tx.tmp \
-    --certificate deleg.cert
+    --out-file ${SP_NODE_ID}/tx.tmp \
+    --certificate ${SP_NODE_ID}/deleg.cert
 
   fee=$(cardano-cli transaction calculate-min-fee \
-    --tx-body-file tx.tmp \
+    --tx-body-file ${SP_NODE_ID}/tx.tmp \
     --tx-in-count 1 \
     --tx-out-count 1 \
     --testnet-magic 888 \
     --witness-count 2 \
     --byron-witness-count 0 \
-    --protocol-params-file ./params.json | awk '{ print $1 }')
+    --protocol-params-file ${SP_NODE_ID}/params.json | awk '{ print $1 }')
 
   initialBalance=$(getAddressBalance "$paymentAddr")
   txOut=$((initialBalance - fee))
@@ -420,18 +426,18 @@ deregisterPool() {
     --tx-out "$paymentAddr"+"$txOut" \
     --invalid-hereafter 5000000 \
     --fee "$fee" \
-    --certificate-file deleg.cert \
-    --out-file tx.raw
+    --certificate-file ${SP_NODE_ID}/deleg.cert \
+    --out-file ${SP_NODE_ID}/tx.raw
 
   cardano-cli transaction sign \
-    --tx-body-file tx.raw \
+    --tx-body-file ${SP_NODE_ID}/tx.raw \
     --signing-key-file "$delegatorPaymentSKey" \
     --signing-key-file "$delegatorStakeSKey" \
     --testnet-magic 888 \
-    --out-file tx.signed
+    --out-file ${SP_NODE_ID}/tx.signed
 
   cardano-cli transaction submit \
-    --tx-file tx.signed \
+    --tx-file ${SP_NODE_ID}/tx.signed \
     --testnet-magic 888
 
   updatedBalance=$(getAddressBalance "$paymentAddr")
@@ -448,7 +454,7 @@ deregisterPool() {
   cardano-cli stake-pool deregistration-certificate \
     --cold-verification-key-file "$coldVKey" \
     --epoch "$RETIRING_EPOCH" \
-    --out-file pool.dereg
+    --out-file ${SP_NODE_ID}/pool.dereg
 
   utxo=$(cardano-cli query utxo --address "$genesisAddr" --testnet-magic 888 | awk 'NR == 3 {printf("%s#%s", $1, $2)}')
 
@@ -457,17 +463,17 @@ deregisterPool() {
     --tx-out "$genesisAddr"+0 \
     --invalid-hereafter 500000 \
     --fee 0 \
-    --certificate-file pool.dereg \
-    --out-file tx.tmp
+    --certificate-file ${SP_NODE_ID}/pool.dereg \
+    --out-file ${SP_NODE_ID}/tx.tmp
 
   fee=$(cardano-cli transaction calculate-min-fee \
-    --tx-body-file tx.tmp \
+    --tx-body-file ${SP_NODE_ID}/tx.tmp \
     --tx-in-count 1 \
     --tx-out-count 1 \
     --testnet-magic 888 \
     --witness-count 3 \
     --byron-witness-count 0 \
-    --protocol-params-file ./params.json | awk '{ print $1 }')
+    --protocol-params-file ${SP_NODE_ID}/params.json | awk '{ print $1 }')
 
   initialBalance=$(getAddressBalance "$genesisAddr")
   txOut=$((initialBalance - fee))
@@ -477,19 +483,19 @@ deregisterPool() {
     --tx-out "$genesisAddr"+"$txOut" \
     --invalid-hereafter 500000 \
     --fee "$fee" \
-    --certificate-file pool.dereg \
-    --out-file tx.raw
+    --certificate-file ${SP_NODE_ID}/pool.dereg \
+    --out-file ${SP_NODE_ID}/tx.raw
 
   cardano-cli transaction sign \
-    --tx-body-file tx.raw \
+    --tx-body-file ${SP_NODE_ID}/tx.raw \
     --signing-key-file "$genesisSKey" \
     --signing-key-file "$coldKey" \
     --signing-key-file "$stakeKey" \
     --testnet-magic 888 \
-    --out-file tx.signed
+    --out-file ${SP_NODE_ID}/tx.signed
 
   cardano-cli transaction submit \
-    --tx-file tx.signed \
+    --tx-file ${SP_NODE_ID}/tx.signed \
     --testnet-magic 888
 
   updatedBalance=$(getAddressBalance "$genesisAddr")