+++ packages/tx-construction/test/tx-builder/TxBuilderDelegatePortfolio.test.ts
@@ -12,6 +12,7 @@ import {
 } from '../../src';
 import { GreedyInputSelector, GreedySelectorProps, roundRobinRandomImprove } from '@cardano-sdk/input-selection';
 import { dummyLogger } from 'ts-log';
+import { mockTxEvaluator } from './mocks';
 import { mockProviders as mocks } from '@cardano-sdk/util-dev';
 import uniqBy from 'lodash/uniqBy';
 
@@ -112,6 +113,7 @@ const createTxBuilder = async ({
       logger: dummyLogger,
       outputValidator,
       txBuilderProviders,
+      txEvaluator: mockTxEvaluator,
       witnesser: util.createBip32Ed25519Witnesser(asyncKeyAgent)
     }),
     txBuilderProviders,
@@ -120,6 +122,7 @@ const createTxBuilder = async ({
       logger: dummyLogger,
       outputValidator,
       txBuilderProviders,
+      txEvaluator: mockTxEvaluator,
       witnesser: util.createBip32Ed25519Witnesser(asyncKeyAgent)
     })
   };
@@ -306,6 +309,11 @@ describe('TxBuilder/delegatePortfolio', () => {
   describe('pre-existing multi-delegation on all stake keys', () => {
     let txBuilderProviders: jest.Mocked<TxBuilderProviders>;
     let nonDelegatingWalletTxBuilder: GenericTxBuilder;
+    let multiDelegatingWalletTxBuilder: GenericTxBuilder;
+    let singleDelegatingWalletTxBuilder: GenericTxBuilder;
+    let singleDelegatingWalletGroupedAddresses: GroupedAddress[];
+    let multiDelegatingWalletGroupedAddresses: GroupedAddress[];
+
     const deposit = 5n;
 
     beforeEach(async () => {
@@ -329,11 +337,60 @@ describe('TxBuilder/delegatePortfolio', () => {
           ]
         })
       ).txBuilder;
+
+      const multiDelegatingWalletTxBuilderFactory = await createTxBuilder({
+        keyAgent,
+        stakeDelegations: [
+          { credentialStatus: Cardano.StakeCredentialStatus.Registered, poolId: poolIds[0] },
+          { credentialStatus: Cardano.StakeCredentialStatus.Registered, poolId: poolIds[1] }
+        ]
+      });
+
+      multiDelegatingWalletTxBuilder = multiDelegatingWalletTxBuilderFactory.txBuilder;
+      multiDelegatingWalletGroupedAddresses = multiDelegatingWalletTxBuilderFactory.groupedAddresses;
+
+      const singleDelegatingWalletTxBuilderFactory = await createTxBuilder({
+        keyAgent,
+        stakeDelegations: [{ credentialStatus: Cardano.StakeCredentialStatus.Registered, poolId: poolIds[0] }]
+      });
+
+      singleDelegatingWalletTxBuilder = singleDelegatingWalletTxBuilderFactory.txBuilder;
+      singleDelegatingWalletGroupedAddresses = singleDelegatingWalletTxBuilderFactory.groupedAddresses;
     });
 
     it('does nothing and uses random improve input selector when the wallet is not delegating and delegatePortfolio is given a null portfolio', async () => {
       const tx = await nonDelegatingWalletTxBuilder.delegatePortfolio(null).build().inspect();
       expect(tx.body.certificates?.length).toBe(0);
+      expect(GreedyInputSelector).not.toHaveBeenCalled();
+      expect(roundRobinRandomImprove).toHaveBeenCalled();
+    });
+
+    it('does nothing and uses random improve input selector when the wallet is multi delegating and transaction is not changing delegation', async () => {
+      const tx = await multiDelegatingWalletTxBuilder
+        .addOutput(
+          multiDelegatingWalletTxBuilder
+            .buildOutput({ address: multiDelegatingWalletGroupedAddresses[0].address, value: { coins: 10n } })
+            .toTxOut()
+        )
+        .build()
+        .inspect();
+      expect(tx.body.certificates).toBeUndefined();
+      expect(tx.body.outputs.length).toBeTruthy();
+      expect(GreedyInputSelector).not.toHaveBeenCalled();
+      expect(roundRobinRandomImprove).toHaveBeenCalled();
+    });
+
+    it('does nothing and uses random improve input selector when the wallet is single delegating and transaction is not changing delegation', async () => {
+      const tx = await singleDelegatingWalletTxBuilder
+        .addOutput(
+          singleDelegatingWalletTxBuilder
+            .buildOutput({ address: singleDelegatingWalletGroupedAddresses[0].address, value: { coins: 10n } })
+            .toTxOut()
+        )
+        .build()
+        .inspect();
+      expect(tx.body.certificates).toBeUndefined();
+      expect(tx.body.outputs.length).toBeTruthy();
       expect(GreedyInputSelector).not.toHaveBeenCalled();
       expect(roundRobinRandomImprove).toHaveBeenCalled();
     });